size_t instead of int

+++ modules/dsp/chowdsp_math/Math/chowdsp_ChebyshevPolynomials.h
@@ -5,7 +5,7 @@ namespace chowdsp
 #ifndef DOXYGEN
 namespace cheby_detail
 {
-    template <typename T, int max_order = 32>
+    template <typename T, size_t max_order = 32>
     constexpr auto cheby0()
     {
         Polynomial<T, max_order, poly_order_ascending> poly {};
@@ -13,7 +13,7 @@ namespace cheby_detail
         return poly;
     }
 
-    template <typename T, int max_order = 32>
+    template <typename T, size_t max_order = 32>
     constexpr auto cheby1()
     {
         Polynomial<T, max_order, poly_order_ascending> poly {};
@@ -24,19 +24,19 @@ namespace cheby_detail
 #endif
 
 /** Recursively computes a Chebyshev polynomial given two previous Chebyshev polynomials */
-template <typename T, int max_order = 32>
+template <typename T, size_t max_order = 32>
 constexpr auto chebyshev_polynomial_recurse (const Polynomial<T, max_order, poly_order_ascending>& cheby_n1,
                                              const Polynomial<T, max_order, poly_order_ascending>& cheby_n2)
 {
     Polynomial<T, max_order, poly_order_ascending> cheby_n {};
     cheby_n.coeffs[0] = -cheby_n2.coeffs[0];
-    for (size_t i = 1; i < (size_t) max_order; ++i)
+    for (size_t i = 1; i < max_order; ++i)
         cheby_n.coeffs[i] = (T) 2 * cheby_n1.coeffs[i - 1] - cheby_n2.coeffs[i];
     return cheby_n;
 }
 
 /** Computes a Chebyshev polynomial of a given order. */
-template <typename T, int order, int max_order = 32>
+template <typename T, size_t order, size_t max_order = 32>
 constexpr auto chebyshev_polynomial()
 {
     if constexpr (order == 0)
@@ -48,7 +48,7 @@ constexpr auto chebyshev_polynomial()
     auto cheby_n2 = cheby_detail::cheby0<T, 32>();
     auto cheby_n1 = cheby_detail::cheby1<T, 32>();
     auto cheby_n = Polynomial<T, max_order, poly_order_ascending> {};
-    for (int i = 2; i <= order; ++i)
+    for (size_t i = 2; i <= order; ++i)
     {
         cheby_n = chebyshev_polynomial_recurse<T, max_order> (cheby_n1, cheby_n2);
         cheby_n2 = cheby_n1;
